pyproject.toml
[tool.poetry]
name = "qa-api-project"
version = "0.1.0"
description = ""
authors = ["i.lobusov <i.lobusov@bi.zone>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
ruff = "^0.6.1"
pytest = "^8.3.2"
yarl = "^1.9.4"
requests = "^2.32.3"
backoff = "^2.2.1"
pydantic = "^2.8.2"
pydantic-settings = "^2.4.0"
faker = "^27.0.0"
polyfactory = "^2.16.2"
pytest-xdist = "^3.6.1"
allure-pytest = "^2.13.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
